Rename pammap.hpp to core.hpp
