+++ pammap/core/core.hpp
